% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_bens_cargo.R
\name{baixar_bens_cargo}
\alias{baixar_bens_cargo}
\title{baixa bens dos candidatos por cargo}
\usage{
baixar_bens_cargo(ano, uf, cod_cargo)
}
\arguments{
\item{ano}{ano da eleicao}

\item{uf}{uf dos candidatos}

\item{cod_cargo}{codigo do cargo dos candidatos}
}
\value{

}
\description{
baixa bens dos candidatos por cargo
}
