% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_bens_cargo_BR.R
\name{baixar_bens_cargo_BR}
\alias{baixar_bens_cargo_BR}
\title{baixa bens de todos os candidatos do cargo desejado no brasil}
\usage{
baixar_bens_cargo_BR(ano, cod_cargo)
}
\arguments{
\item{ano}{ano da eleicao}

\item{cod_cargo}{codigo do cargo}
}
\value{

}
\description{
baixa bens de todos os candidatos do cargo desejado no brasil
}
