% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captura_bens.R
\name{captura_bens}
\alias{captura_bens}
\title{captura_bens}
\usage{
captura_bens(id_candidato, uf = "PE", ano = 2022)
}
\arguments{
\item{id_candidato}{string de ids do candidato}

\item{uf}{uf do candidato}

\item{ano}{ano da eleicao}
}
\value{

}
\description{
captura_bens
}
