% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compara_votos_turno.R
\name{compara_votos_turno}
\alias{compara_votos_turno}
\title{compara_votos_turno}
\usage{
compara_votos_turno(cargo = "PRESIDENTE")
}
\arguments{
\item{cargo}{cargo da disputa}
}
\value{
df comparando os votos no 1 e no 2 turno
}
\description{
compara_votos_turno
}
