% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_info_cand.R
\name{pega_info_cand}
\alias{pega_info_cand}
\title{pega info candidatos}
\usage{
pega_info_cand(ano = 2022, sigla_estado = "PE", cod_cargo = 7)
}
\arguments{
\item{ano}{ano eleicao}

\item{sigla_estado}{sigla uf}

\item{cod_cargo}{codigo cargo}
}
\value{

}
\description{
pega info candidatos
}
