% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_info_cand_BR.R
\name{pega_info_cand_BR}
\alias{pega_info_cand_BR}
\title{pega informacoes dos candidatos do cargo desejado de todo brasil}
\usage{
pega_info_cand_BR(ano, cod_cargo)
}
\arguments{
\item{ano}{ano da eleicao}

\item{cod_cargo}{codigo cargo}
}
\value{

}
\description{
pega informacoes dos candidatos do cargo desejado de todo brasil
}
