% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raspa_votacao_g1.R
\name{raspa_votacao_g1}
\alias{raspa_votacao_g1}
\title{baixa um dataset com os dados de votacao do g1}
\usage{
raspa_votacao_g1(turno = 1, cargo = "GOVERNADOR", uf = "SP")
}
\arguments{
\item{turno}{turno da eleicao}

\item{cargo}{cargo ou governador ou presidente}

\item{uf}{uf desejada}
}
\value{
dataset com dados de votacao do g1
}
\description{
baixa um dataset com os dados de votacao do g1
}
