% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raspa_votacao_g1_br.R
\name{raspa_votacao_g1_br}
\alias{raspa_votacao_g1_br}
\title{baixa um dataset com os dados do brasil de votacao do g1}
\usage{
raspa_votacao_g1_br(turno = 1, cargo = "GOVERNADOR")
}
\arguments{
\item{turno}{turno desejado default 1}

\item{cargo}{cargo desejado default governador}
}
\value{
dataset com dados de votacao do brasil
}
\description{
baixa um dataset com os dados do brasil de votacao do g1
}
\examples{
raspa_votacao_g1_br(tunro = 1, cargo = "GOVERNADOR")
}
