% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_receitas_cand.R
\name{baixa_receitas_cand}
\alias{baixa_receitas_cand}
\title{baixa_receitas_candidatos}
\usage{
baixa_receitas_cand(ano_eleicao, uf, cod_cargo, n_partido, n_cand, cod_cand)
}
\arguments{
\item{ano_eleicao}{ano da eleicao}

\item{uf}{uf desejada}

\item{cod_cargo}{codigo cargo desejado}

\item{n_partido}{n partido}

\item{n_cand}{n_cand}

\item{cod_cand}{cod_cand}
}
\value{

}
\description{
baixa_receitas_candidatos
}
